Refactor lenderVaults components
